src/main/java/com/bookingApp/service/UserService.java
package com.bookingApp.service;

import com.bookingApp.model.User;
import com.bookingApp.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

@Service
public class UserService {

    private final UserRepository userRepository;
    private final PasswordEncoder passwordEncoder;

    @Autowired
    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
    }

    // check registerUser 4roles
    public User registerUser(String username, String password, String role) {
        // create user
        User user = new User();
        user.setUsername(username);
        user.setPassword(passwordEncoder.encode(password));
        user.setRole(role);  // set strin role

        // save in db
        return userRepository.save(user);
    }
}

